move common code for intensity blackbody and calculate_p
